tests/test_template.py
"""
Comprehensive Test Template

This file provides a comprehensive template for writing unit tests for the project.
It includes best practices, common patterns, and examples for different types of tests.

Copy this file and modify it according to your testing needs.
"""

import unittest
import sys
import os
from unittest.mock import Mock, patch
from pathlib import Path
import tempfile
import shutil
import pytest
from src import main

# Add the src directory to the path so we can import modules
sys.path.insert(0, str(Path(__file__).parent.parent / "src"))


class TestTemplate(unittest.TestCase):
    """Template class for unit tests with comprehensive examples"""

    def setUp(self):
        """Set up test fixtures before each test method"""
        # Initialize test data and mock objects here
        self.test_data = "sample data"
        self.expected_result = "expected output"
        self.mock_object = Mock()

    def tearDown(self):
        """Clean up after each test method"""
        # Clean up resources here
        # Reset mocks
        self.mock_object.reset_mock()

    def test_sample_functionality(self):
        """Test basic functionality"""
        # Arrange

        # Act
        # result = my_function(input_data)

        # Assert
        # self.assertEqual(result, self.expected_result)

    def test_with_mocking(self):
        """Example of testing with mocks"""
        # Arrange
        self.mock_object.some_method.return_value = "mocked result"

        # Act
        result = self.mock_object.some_method()

        # Assert
        self.assertEqual(result, "mocked result")
        self.mock_object.some_method.assert_called_once()

    @patch("src.main.sanitize_input")
    def test_with_patch_decorator(self, mock_function):
        """Example of testing with patch decorator"""
        # Arrange
        mock_function.return_value = "patched result"

        # Act
        result = main.sanitize_input("input")

        # Assert
        self.assertEqual(result, "patched result")
        mock_function.assert_called_once_with("input")

    def test_context_manager_mocking(self):
        """Example of testing with context manager mocking"""
        with patch("src.main.sanitize_input") as mock_function:
            # Arrange
            mock_function.return_value = "sanitized input"

            # Act
            result = main.sanitize_input("test input")

            # Assert
            self.assertEqual(result, "sanitized input")
            mock_function.assert_called_once_with("test input")

    def test_edge_cases(self):
        """Test edge cases and error conditions"""
        # Test with empty input
        # with self.assertRaises(ValueError):
        #     my_function("")

        # Test with None input
        # with self.assertRaises(TypeError):
        #     my_function(None)

        # Test with boundary values
        # result = my_function(0)
        # self.assertEqual(result, expected_boundary_result)

    def test_error_handling(self):
        """Test error handling scenarios"""
        # Test expected exceptions
        with self.assertRaises(ValueError) as context:
            # Code that should raise ValueError
            raise ValueError("Test error")

        self.assertIn("Test error", str(context.exception))

    def test_data_driven(self):
        """Example of data-driven testing"""
        test_cases = [
            ("input1", "expected1"),
            ("input2", "expected2"),
            ("input3", "expected3"),
        ]

        for input_val, _ in test_cases:
            with self.subTest(input=input_val):
                # Act
                # result = my_function(input_val)

                # Assert
                # self.assertEqual(result, _)
                pass

    def test_assertion_methods(self):
        """Demonstrate various assertion methods"""
        # Basic assertions
        self.assertEqual(1 + 1, 2)
        self.assertNotEqual(1, 2)
        self.assertIsNone(None)
        self.assertIsNotNone("not none")
        self.assertIn(1, [1, 2, 3])
        self.assertNotIn(4, [1, 2, 3])

        # String assertions
        self.assertIn("lo", "hello")

        # List/Dict assertions
        self.assertEqual([1, 2], [1, 2])
        self.assertDictEqual({"a": 1}, {"a": 1})

        # Exception assertions
        with self.assertRaises(ValueError):
            raise ValueError()

    def test_integration(self):
        """Test integration with other components"""
        # Test how this component interacts with others
        # This might involve setting up multiple mocks or real objects


class TestAsyncFunctionality:
    """Template for testing async functions"""

    @pytest.mark.asyncio
    async def test_async_function(self):
        """Test async function"""
        # Arrange

        # Act
        # result = await async_function(input_data)

        # Assert
        # assert result == expected_result

    @pytest.mark.asyncio
    async def test_async_with_mock(self):
        """Test async function with mocking"""
        # Arrange
        # with patch("actual.module.path") as mock_async_func:
        #     mock_async_func.return_value = "async mock result"

        # Act
        # result = await function_that_calls_async(mock_async_func)

        # Assert
        # assert result == "async mock result"
        pass  # Placeholder test - implement when needed


class TestClassWithSetup(unittest.TestCase):
    """Example test class with more complex setup"""

    @classmethod
    def setUpClass(cls):
        """Set up class-level fixtures once for all tests"""
        cls.shared_data = "shared across all tests"
        cls.expensive_resource = Mock()  # Simulate expensive resource

    @classmethod
    def tearDownClass(cls):
        """Clean up class-level fixtures"""
        # Clean up expensive resources

    def setUp(self):
        """Set up test fixtures before each test method"""
        self.test_specific_data = "unique to each test"

    def tearDown(self):
        """Clean up after each test method"""
        # Reset any test-specific state

    def test_using_class_setup(self):
        """Test that uses class-level setup"""
        self.assertEqual(self.shared_data, "shared across all tests")
        self.assertEqual(self.test_specific_data, "unique to each test")


class TestFileOperations(unittest.TestCase):
    """Template for testing file operations"""

    def setUp(self):
        """Create temporary files/directories for testing"""
        self.temp_dir = tempfile.mkdtemp()
        self.temp_file = os.path.join(self.temp_dir, "test.txt")

    def tearDown(self):
        """Clean up temporary files"""
        shutil.rmtree(self.temp_dir)

    def test_file_creation(self):
        """Test file creation"""
        # Create a test file
        with open(self.temp_file, "w", encoding="utf-8") as f:
            f.write("test content")

        # Verify file exists
        self.assertTrue(os.path.exists(self.temp_file))

        # Read and verify content
        with open(self.temp_file, "r", encoding="utf-8") as f:
            content = f.read()
        self.assertEqual(content, "test content")


class TestDatabaseOperations(unittest.TestCase):
    """Template for testing database operations"""

    def setUp(self):
        """Set up database connection/mock"""
        # self.db_connection = create_test_database()
        # Or use mocking
        self.db_mock = Mock()

    def tearDown(self):
        """Clean up database"""
        # close_test_database(self.db_connection)

    def test_database_query(self):
        """Test database query"""
        # Mock database response
        self.db_mock.query.return_value = [{"id": 1, "name": "test"}]

        # Act
        # result = get_data_from_db(self.db_mock)

        # Assert
        # self.assertEqual(len(result), 1)
        # self.assertEqual(result[0]["name"], "test")


# Example of parameterized tests (Python 3.4+)
class TestParameterized(unittest.TestCase):
    """Example of parameterized tests"""

    def test_parameterized_example(self):
        """Example of manual parameterization"""
        for param, _ in [("a", 1), ("b", 2), ("c", 3)]:
            with self.subTest(param=param):
                # result = process_param(param)
                # self.assertEqual(result, _)
                pass


if __name__ == "__main__":
    # Run the tests
    unittest.main(verbosity=2)

    # Alternative: Run with coverage
    # coverage run -m unittest discover
    # coverage report

    # Alternative: Run specific test
    # python -m unittest test_template.TestTemplate.test_sample_functionality

    # Alternative: Run with test discovery
    # python -m unittest discover -s . -p "test_*.py"
