pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-template"
version = "0.1.0"
description = "A project template"
authors = [
    {name = "Author Name", email = "author@example.com"}
]
dependencies = [
    "typer",
    "python-dotenv",
    "pytest-asyncio"
]

[tool.setuptools]
packages = ["find:"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = ["src"]